feat: Added Card component, refactored Popular, Results
Card component receives 5 string props: header,
subheader (optional), avatar, href and name.
In addition, the unordered list is passed as a
React children.

It displays a small div, with the following
HTML elements: h4, img, optional h4,
h2 and anchor. Plus, the ul (children)
is appended.

Results component is refactored to use 2 cards:
one for the winner and one for the loser.

Popular component is refactored to
use the Card component.