feat: Added Reset user functionality
Added Reset Button on Results component,
and onReset function on Battle component.

The UI is on the Results component,
but the onReset function is on Battle component.

onReset reset the state to both players
being null and battle to false, to
create a re render of Battle component
to its initial state.

Also, propTypes are added to Results
components: 2 strings and the onReset
callback function.